# [Need test] vm locking refactoring