--- a/rpcs3/Emu/Memory/vm_locking.h
+++ b/rpcs3/Emu/Memory/vm_locking.h
@@ -7,8 +7,6 @@ class shared_mutex;
 
 namespace vm
 {
-	extern shared_mutex g_mutex;
-
 	extern thread_local atomic_t<cpu_thread*>* g_tls_locked;
 
 	enum range_lock_flags : u64
@@ -110,24 +108,12 @@ namespace vm
 	void temporary_unlock(cpu_thread& cpu) noexcept;
 	void temporary_unlock() noexcept;
 
-	class reader_lock final
-	{
-		bool m_upgraded = false;
-
-	public:
-		reader_lock(const reader_lock&) = delete;
-		reader_lock& operator=(const reader_lock&) = delete;
-		reader_lock();
-		~reader_lock();
-
-		void upgrade();
-	};
-
 	struct writer_lock final
 	{
 		writer_lock(const writer_lock&) = delete;
 		writer_lock& operator=(const writer_lock&) = delete;
-		writer_lock(u32 addr = 0);
+		writer_lock();
+		writer_lock(u32 addr, u32 size = 0, u64 flags = range_locked);
 		~writer_lock();
 	};
 } // namespace vm