The idea is simple, since g_mutex and g_range_lock usually go together, simplify locking mechanics and potentially make them more efficient.