@AniLeo 2022-04-30
Tested a few games to check the stability: 
- Red Dead Redemption
- Yakuza: Dead Souls
- NieR Replicant
- vsh.self


Seems to work fine, needs performance comparison tests