@kd-11 2022-04-30
Quick testing shows massive performance improvements in SPU-limited games.
MGS4 52->72 fps average (+40%)
Sonic Unleashed 41 -> 46 fps (+12%)
I'm checking more titles.