@Xcedf 2022-04-30
SPU Test some improvements

Master:
SPU Test v1.1.2 by GalCiv
SPU Task Avalanche completed in 5239 ms (PS3: 2740 ms)
PPU/SPU Ping-Pong completed in 1549 ms (PS3: 3045 ms)
SPU MFC 64 Bits War completed in 4589 ms (PS3: 3370 ms)
PPU/SPU MFC 64 Bits War completed in 6397 ms (PS3: 4443 ms)
SPU Integer Perf completed in 1225 ms (PS3: 8666 ms)
SPU Float Perf completed in 324 ms (PS3: 2379 ms)
SPU SpinLock completed in 3605 ms (PS3: 4409 ms)
PUTLLUC Perf completed in 8072 ms (PS3: 3853 ms)
PUTLLC Perf completed in 1500 ms (PS3: 3364 ms)
PUT Perf completed in 1564 ms (PS3: 3984 ms)
Large PUT Perf completed in 1239 ms (PS3: 4454 ms)
--Completed 11 tests in 35305 ms--

PR:
SPU Test v1.1.2 by GalCiv
SPU Task Avalanche completed in 2704 ms (PS3: 2740 ms)
PPU/SPU Ping-Pong completed in 1293 ms (PS3: 3045 ms)
SPU MFC 64 Bits War completed in 4405 ms (PS3: 3370 ms)
PPU/SPU MFC 64 Bits War completed in 7038 ms (PS3: 4443 ms)
SPU Integer Perf completed in 1223 ms (PS3: 8666 ms)
SPU Float Perf completed in 324 ms (PS3: 2379 ms)
SPU SpinLock completed in 3867 ms (PS3: 4409 ms)
PUTLLUC Perf completed in 7629 ms (PS3: 3853 ms)
PUTLLC Perf completed in 2322 ms (PS3: 3364 ms)
PUT Perf completed in 1534 ms (PS3: 3984 ms)
Large PUT Perf completed in 1237 ms (PS3: 4454 ms)
--Completed 11 tests in 33578 ms--