@AniLeo 2022-04-30
+22 average FPS in Sonic Unleashed in-engine cutscene on 5900HX

**Master**
![Screenshot from 2022-04-30 16-52-15](https://user-images.githubusercontent.com/10283761/166112976-50bf751f-3ef8-4b93-9953-5818aabba3e4.png)

**PR**
![Screenshot from 2022-04-30 16-55-21](https://user-images.githubusercontent.com/10283761/166112981-3c6d4ddb-5a9a-4de3-abe2-60dddb6eb243.png)