@AniLeo 2022-04-30
Drakengard 3 is around 20 FPS slower on PR for some reason, tried twice on each build and got consistent results

**Master**
![Screenshot from 2022-04-30 17-25-12](https://user-images.githubusercontent.com/10283761/166114390-34557524-fa04-4877-b670-629c7e64963d.png)

**PR**
![Screenshot from 2022-04-30 17-28-24](https://user-images.githubusercontent.com/10283761/166114400-40b99631-b35a-470e-96e8-1fdc8d985d22.png)