@Asinin3 2022-04-30
5950X - Keep in mind many other CPU's will see more gains than I did, or possibly gains where I saw none depending on bottleneck.
Metal Gear Solid 4: 30fps > 60fps lock (fps capped)
Persona 5: 75 > 78 FPS (Central Street) 
Sonic Unleashed: 53 > 56 (starting spot)
Red Dead Redemption: No change (seems same on intel)
Killzone 3: No change
R&C Quest for Booty: No change
Infamous: No change