@AniLeo 2022-04-30
+13 average FPS in Drakengard 3 in this scene that is not RSX bottlenecked

**Master**
![Screenshot from 2022-04-30 18-01-45](https://user-images.githubusercontent.com/10283761/166115220-90ff4de0-4c5e-41e6-b688-2b4af4ce5931.png)

**PR**
![Screenshot from 2022-04-30 17-58-19](https://user-images.githubusercontent.com/10283761/166115217-7c5468a3-a29e-424e-b2e7-cae7bb6ea5e3.png)