@FlexBy420 2022-04-30
Ryzen 5 3600, 
very small gain in Uncharted 3: Drake's Deception
no gain at all in Pirates of the Caribbean: At World's End

Master
![obraz_2022-04-30_191941524](https://user-images.githubusercontent.com/68403300/166115745-f75d1bae-1760-4aef-bb97-68a528740db4.png)
![image](https://user-images.githubusercontent.com/68403300/166116003-9701592d-98f0-49e2-9ba6-5dcfadff23c2.png)

PR
![image](https://user-images.githubusercontent.com/68403300/166115765-2f2fbc17-52a7-4181-a3cd-7cbe424e7afe.png)
![image](https://user-images.githubusercontent.com/68403300/166116009-52d81e0c-186e-4a11-8850-6aebb0338276.png)