@AniLeo 2022-04-30
+10 average FPS (+27%) on Metal Gear Solid 4 in Red Zone on 5900HX

**Master**
![Screenshot from 2022-04-30 18-53-52](https://user-images.githubusercontent.com/10283761/166116911-91211b03-4827-44c0-b3ea-1a29629cef43.png)

**PR**
![Screenshot from 2022-04-30 18-51-09](https://user-images.githubusercontent.com/10283761/166116914-e94df7a0-8cbc-42ea-8f47-264f269e4e02.png)