@cipherxof 2022-04-30
Definitely seeing a slight performance improvement here in MGS4. Unfortunately, it's still not up to par with custom builds almost entirely because of the zcull hack that those include.

**Master (relaxed zcull):**

![master](https://user-images.githubusercontent.com/5994581/166118682-0bf27a0d-1888-4277-a427-a92349b19673.png)

![master2](https://user-images.githubusercontent.com/5994581/166118688-6098c578-da18-4835-83b7-765033191524.png)

**PR (relaxed zcull):**

![pr](https://user-images.githubusercontent.com/5994581/166118693-3562a9fa-cff2-4cef-bc5f-54bf047c81ea.png)

![pr2](https://user-images.githubusercontent.com/5994581/166118697-540542f7-5bda-4917-8879-65750632aba4.png)

**PR (zcull hack):**

![custom2](https://user-images.githubusercontent.com/5994581/166118706-6c8b125e-9467-4a2a-8db8-a685c6b9702a.png)

![custom](https://user-images.githubusercontent.com/5994581/166118710-bf832fb4-f097-45b1-8732-d02d03553e79.png)