@Hasster1 2022-04-30
It's not so great for me...
i5-4210u
Both HAZE and InFamous are stuttering on this pr (Infamous with disabled video output: master - 12-13 fps, while this pr gives me 8-12, Haze is stuttering as well.)
But Serious Sam 3 actually gives me more fps than master (24-25 on master, 26-28 on this pr ) without any stuttering. 

Is it because of my craptop?