@Asinin3 2022-04-30
@Hasster1 You need to test games with a SPU bottleneck, infamous isn't one of them, and I doubt Haze or Serious Sam are either. test some of the games mentioned here to improve largely and see what you find.