@Hasster1 2022-04-30
It's not about improvement, it's about regression. Even if these games are without SPU bottleneck, they're still affected negatively on my laptop, which isn't good. I don't want to lose any amount of fps that i've got.