@Asinin3 2022-04-30
Are you using RPCS3's CPU Scheduler, or do you have it set to Operating System? Ani found the PR to negatively impact the RPCS3 scheduler, but faster on OS scheduler. It's likely that the RPCS3 scheduler needs to be re-written to work with this change.