@AniLeo 2022-04-30
Just for reference, performance regressions on specifications that are very, very low end (i.e mobile dual core) are not really blockers, that's already waaay below recommended specifications