@solarmystic 2022-04-30
On my **12400F**, SPU Test showed over 10% improvement in completion time. 

Master:-
SPU Test v1.1.2 by GalCiv
SPU Task Avalanche completed in 6221 ms (PS3: 2740 ms)
PPU/SPU Ping-Pong completed in 2252 ms (PS3: 3045 ms)
SPU MFC 64 Bits War completed in 5834 ms (PS3: 3370 ms)
PPU/SPU MFC 64 Bits War completed in 7696 ms (PS3: 4443 ms)
SPU Integer Perf completed in 2027 ms (PS3: 8666 ms)
SPU Float Perf completed in 620 ms (PS3: 2379 ms)
SPU SpinLock completed in 4445 ms (PS3: 4409 ms)
PUTLLUC Perf completed in 10571 ms (PS3: 3853 ms)
PUTLLC Perf completed in 3138 ms (PS3: 3364 ms)
PUT Perf completed in 1609 ms (PS3: 3984 ms)
Large PUT Perf completed in 1723 ms (PS3: 4454 ms)
**--Completed 11 tests in 46139 ms--**

PR:-
SPU Test v1.1.2 by GalCiv
SPU Task Avalanche completed in 3451 ms (PS3: 2740 ms)
PPU/SPU Ping-Pong completed in 1822 ms (PS3: 3045 ms)
SPU MFC 64 Bits War completed in 5591 ms (PS3: 3370 ms)
PPU/SPU MFC 64 Bits War completed in 7132 ms (PS3: 4443 ms)
SPU Integer Perf completed in 1989 ms (PS3: 8666 ms)
SPU Float Perf completed in 552 ms (PS3: 2379 ms)
SPU SpinLock completed in 5067 ms (PS3: 4409 ms)
PUTLLUC Perf completed in 9855 ms (PS3: 3853 ms)
PUTLLC Perf completed in 2716 ms (PS3: 3364 ms)
PUT Perf completed in 1621 ms (PS3: 3984 ms)
Large PUT Perf completed in 1684 ms (PS3: 4454 ms)
**--Completed 11 tests in 41482 ms--**